+++ packages/algob/src/lib/algo-operator.ts
@@ -73,7 +73,7 @@ export interface AlgoOperator {
 		txWriter: txWriter,
 		scTmplParams?: SCParams
 	) => Promise<rtypes.AppInfo>;
-	waitForConfirmation: (txId: string) => Promise<ConfirmedTxInfo>;
+	waitForConfirmation: (txId: string) => Promise<TxnReceipt>;
 	getAssetByID: (assetIndex: number | bigint) => Promise<modelsv2.Asset>;
 	optInAccountToASA: (
 		asaName: string,
@@ -120,7 +120,7 @@ export interface AlgoOperator {
 	sendAndWait: (
 		rawTxns: Uint8Array | Uint8Array[],
 		waitRounds: number
-	) => Promise<ConfirmedTxInfo>;
+	) => Promise<TxnReceipt>;
 	getReceiptTxns: (txns: Transaction[]) => Promise<TxnReceipt[]>;
 }
 
@@ -136,28 +136,33 @@ export class AlgoOperatorImpl implements AlgoOperator {
 	 * Send signed transaction to network and wait for confirmation
 	 * @param rawTxns Signed Transaction(s)
 	 * @param waitRounds number of rounds to wait for transaction to be confirmed - default is 10
+	 * @returns TxnReceipt which includes confirmed txn response along with txID
 	 */
 	async sendAndWait(
 		rawTxns: Uint8Array | Uint8Array[],
 		waitRounds = wtypes.WAIT_ROUNDS
-	): Promise<ConfirmedTxInfo> {
+	): Promise<TxnReceipt> {
 		const txInfo = await this.algodClient.sendRawTransaction(rawTxns).do();
 		return await this.waitForConfirmation(txInfo.txId, waitRounds);
 	}
 
-	// Source:
-	// https://github.com/algorand/docs/blob/master/examples/assets/v2/javascript/AssetExample.js#L21
-	// Function used to wait for a tx confirmation
+	/**
+	 * Source: https://github.com/algorand/docs/blob/master/examples/assets/v2/javascript/AssetExample.js#L21
+	 * Function used to wait for a tx confirmation
+	 * @param txId txn ID for which confirmation is required 
+	 * @param waitRounds number of rounds to wait for transaction to be confirmed - default is 10
+	 * @returns TxnReceipt which includes confirmed txn response along with txID
+	 */
 	async waitForConfirmation(
 		txId: string,
 		waitRounds = wtypes.WAIT_ROUNDS
-	): Promise<ConfirmedTxInfo> {
+	): Promise<TxnReceipt> {
 		const pendingInfo = await algosdk.waitForConfirmation(this.algodClient, txId, waitRounds);
 		if (pendingInfo["pool-error"]) {
 			throw new Error(`Transaction Pool Error: ${pendingInfo["pool-error"] as string}`);
 		}
 		if (pendingInfo[confirmedRound] !== null && pendingInfo[confirmedRound] > 0) {
-			return pendingInfo as ConfirmedTxInfo;
+			return { txID: txId, ...pendingInfo as ConfirmedTxInfo } as TxnReceipt
 		}
 		throw new Error("timeout");
 	}