+++ packages/algob/src/types.ts
@@ -486,13 +486,14 @@ export interface Deployer {
 
 	registerSSCInfo: (name: string, sscInfo: rtypes.AppInfo) => void;
 
-	logTx: (message: string, txConfirmation: ConfirmedTxInfo) => void;
+	logTx: (message: string, txConfirmation: TxnReceipt) => void;
 
 	/**
 	 * Send signed transaction to network and wait for confirmation
 	 * @param rawTxns Signed Transaction(s)
+	 * @returns TxnReceipt which includes confirmed txn response along with txID
 	 */
-	sendAndWait: (rawTxns: Uint8Array | Uint8Array[]) => Promise<ConfirmedTxInfo>;
+	sendAndWait: (rawTxns: Uint8Array | Uint8Array[]) => Promise<TxnReceipt>;
 
 	/**
 	 * Return receipts for each transaction in group txn
@@ -605,7 +606,7 @@ export interface Deployer {
 
 	/**
 	 * Queries blockchain for a given transaction and waits until it will be processed. */
-	waitForConfirmation: (txId: string) => Promise<ConfirmedTxInfo>;
+	waitForConfirmation: (txId: string) => Promise<TxnReceipt>;
 
 	/**
 	 * Queries blockchain using algodv2 for asset information by index  */